Clear code add checkstyle.xml
