+++ src/main/java/com/sniklz/pizza/PizzaServiceRenewedApplication.java
@@ -1,0 +1,13 @@
+package com.sniklz.pizza;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PizzaServiceRenewedApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PizzaServiceRenewedApplication.class, args);
+    }
+
+}
