+++ src/main/java/com/sniklz/pizza/dto/response/IngredientResponseDto.java
@@ -1,0 +1,11 @@
+package com.sniklz.pizza.dto.response;
+
+import java.math.BigDecimal;
+import lombok.Data;
+
+@Data
+public class IngredientResponseDto {
+    private long id;
+    private String name;
+    private BigDecimal cost;
+}
