+++ src/main/java/com/sniklz/pizza/model/Pizza.java
@@ -1,0 +1,26 @@
+package com.sniklz.pizza.model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.Table;
+import java.math.BigDecimal;
+import java.util.Set;
+import lombok.Data;
+
+@Entity
+@Table(name = "pizzas")
+@Data
+public class Pizza {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private String description;
+    @ManyToMany
+    private Set<Ingredient> ingredients;
+    private BigDecimal cost;
+}
