+++ src/main/java/com/sniklz/pizza/repository/IngredientRepository.java
@@ -1,0 +1,9 @@
+package com.sniklz.pizza.repository;
+
+import com.sniklz.pizza.model.Ingredient;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
+}
