+++ src/main/java/com/sniklz/pizza/repository/specification/SpecificationManager.java
@@ -1,0 +1,7 @@
+package com.sniklz.pizza.repository.specification;
+
+import org.springframework.data.jpa.domain.Specification;
+
+public interface SpecificationManager<T> {
+    Specification<T> get(String filterKey, String[] params);
+}
