+++ src/main/java/com/sniklz/pizza/repository/specification/SpecificationProvider.java
@@ -1,0 +1,9 @@
+package com.sniklz.pizza.repository.specification;
+
+import org.springframework.data.jpa.domain.Specification;
+
+public interface SpecificationProvider<T> {
+    Specification<T> getSpecification(String[] params);
+
+    String getFilterKey();
+}
