+++ src/main/java/com/sniklz/pizza/service/DefaultService.java
@@ -1,0 +1,10 @@
+package com.sniklz.pizza.service;
+
+public interface DefaultService<T> {
+
+    T save(T model);
+
+    T get(Long id);
+
+    T update(Long id, T model);
+}
