+++ src/main/java/com/sniklz/pizza/service/IngredientService.java
@@ -1,0 +1,8 @@
+package com.sniklz.pizza.service;
+
+import com.sniklz.pizza.model.Ingredient;
+import java.util.List;
+
+public interface IngredientService extends DefaultService<Ingredient> {
+    List<Ingredient> findAll();
+}
