+++ src/main/java/com/sniklz/pizza/service/mapper/RequestMapper.java
@@ -1,0 +1,6 @@
+package com.sniklz.pizza.service.mapper;
+
+public interface RequestMapper<M, R> {
+
+    M toModel(R requestDto);
+}
