+++ src/main/java/com/sniklz/pizza/service/mapper/ResponseMapper.java
@@ -1,0 +1,5 @@
+package com.sniklz.pizza.service.mapper;
+
+public interface ResponseMapper<M, R> {
+    R toDto(M model);
+}
