+++ src/main/java/com/sniklz/pizza/service/mapper/UniversalMapper.java
@@ -1,0 +1,6 @@
+package com.sniklz.pizza.service.mapper;
+
+public interface UniversalMapper<M, R, O>
+        extends RequestMapper<M, R>,
+        ResponseMapper<M, O> {
+}
