--- src/main/java/com/sniklz/pizzaservicerenewed/PizzaServiceRenewedApplication.java
@@ -1,13 +1,0 @@
-package com.sniklz.pizzaservicerenewed;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PizzaServiceRenewedApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PizzaServiceRenewedApplication.class, args);
-    }
-
-}
