--- src/main/java/com/sniklz/pizzaservicerenewed/dto/request/IngredientRequestDto.java
@@ -1,11 +1,0 @@
-package com.sniklz.pizzaservicerenewed.dto.request;
-
-import lombok.Data;
-
-import java.math.BigDecimal;
-
-@Data
-public class IngredientRequestDto {
-    private String name;
-    private BigDecimal cost;
-}
