--- src/main/java/com/sniklz/pizzaservicerenewed/model/Ingredient.java
@@ -1,19 +1,0 @@
-package com.sniklz.pizzaservicerenewed.model;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.math.BigDecimal;
-import java.util.Set;
-
-@Entity
-@Table(name = "ingredients")
-@Data
-public class Ingredient {
-
-    @Id
-    @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private BigDecimal cost;
-}
