--- src/main/java/com/sniklz/pizzaservicerenewed/repository/IngredientRepository.java
@@ -1,12 +1,0 @@
-package com.sniklz.pizzaservicerenewed.repository;
-
-import com.sniklz.pizzaservicerenewed.model.Ingredient;
-import com.sniklz.pizzaservicerenewed.model.Pizza;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Set;
-
-@Repository
-public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
-}
