--- src/main/java/com/sniklz/pizzaservicerenewed/repository/PizzaRepository.java
@@ -1,14 +1,0 @@
-package com.sniklz.pizzaservicerenewed.repository;
-
-import com.sniklz.pizzaservicerenewed.model.Ingredient;
-import com.sniklz.pizzaservicerenewed.model.Pizza;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.stereotype.Repository;
-
-import java.util.Set;
-
-@Repository
-public interface PizzaRepository extends JpaRepository<Pizza, Long>, JpaSpecificationExecutor<Pizza> {
-    //Set<Pizza> getPizzasByIngredients(Set<Ingredient> ingredients);
-}
