--- src/main/java/com/sniklz/pizzaservicerenewed/repository/specification/SpecificationManager.java
@@ -1,7 +1,0 @@
-package com.sniklz.pizzaservicerenewed.repository.specification;
-
-import org.springframework.data.jpa.domain.Specification;
-
-public interface SpecificationManager<T> {
-    Specification<T> get(String filterKey, String[] params);
-}
