--- src/main/java/com/sniklz/pizzaservicerenewed/repository/specification/SpecificationProvider.java
@@ -1,9 +1,0 @@
-package com.sniklz.pizzaservicerenewed.repository.specification;
-
-import org.springframework.data.jpa.domain.Specification;
-
-public interface SpecificationProvider<T> {
-    Specification<T> getSpecification(String[] params);
-
-    String getFilterKey();
-}
