--- src/main/java/com/sniklz/pizzaservicerenewed/service/DefaultService.java
@@ -1,9 +1,0 @@
-package com.sniklz.pizzaservicerenewed.service;
-
-public interface DefaultService<T> {
-
-    T save(T model);
-    T get (Long id);
-    T update(Long id, T model);
-
-}
