--- src/main/java/com/sniklz/pizzaservicerenewed/service/IngredientService.java
@@ -1,13 +1,0 @@
-package com.sniklz.pizzaservicerenewed.service;
-
-import com.sniklz.pizzaservicerenewed.model.Ingredient;
-import com.sniklz.pizzaservicerenewed.model.Pizza;
-
-import java.util.List;
-import java.util.Set;
-
-public interface IngredientService extends DefaultService<Ingredient> {
-    List<Ingredient> findAll();
-
-
-}
