--- src/main/java/com/sniklz/pizzaservicerenewed/service/mapper/RequestMapper.java
@@ -1,6 +1,0 @@
-package com.sniklz.pizzaservicerenewed.service.mapper;
-
-public interface RequestMapper<Model, Request> {
-
-    Model toModel(Request requestDto);
-}
