--- src/main/java/com/sniklz/pizzaservicerenewed/service/mapper/ResponseMapper.java
@@ -1,5 +1,0 @@
-package com.sniklz.pizzaservicerenewed.service.mapper;
-
-public interface ResponseMapper<Model, Response> {
-    Response toDto(Model model);
-}
