--- src/main/java/com/sniklz/pizzaservicerenewed/service/mapper/UniversalMapper.java
@@ -1,6 +1,0 @@
-package com.sniklz.pizzaservicerenewed.service.mapper;
-
-public interface UniversalMapper<Model, RequestDTO, ResponseDTO>
-        extends RequestMapper<Model, RequestDTO>,
-        ResponseMapper<Model, ResponseDTO> {
-}
