+++ src/test/java/com/sniklz/pizza/PizzaServiceRenewedApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sniklz.pizza;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PizzaServiceRenewedApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
