--- src/test/java/com/sniklz/pizzaservicerenewed/PizzaServiceRenewedApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sniklz.pizzaservicerenewed;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PizzaServiceRenewedApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
